+++ src/main.cpp
@@ -21,6 +21,8 @@
 #include "release.h"
 #include "version.h"
 
+#include "NoMenuIconStyle.h"
+
 // define the base class for SingleApplication
 #define QAPPLICATION_CLASS QApplication
 
@@ -197,6 +199,10 @@ int mainStartupMisc(const QStringList &arguments) {
     if (hideMenuIcons) {
         QApplication::setStyle(new NoMenuIconStyle);
     }
+
+#ifdef Q_OS_MAC
+    QApplication::setStyle(new NoMenuIconStyle);
+#endif
 
 
 #ifdef Q_OS_WIN32
